# tsdb: commit data periodically in CreateBlock