To avoid building up data in memory, commit and make a new appender periodically.

The number `commitAfter = 10000` was chosen arbitrarily; testing with 10x more or less gives slightly worse results.

My motivation for looking at this was I could not run `BenchmarkCompaction()` on a VM with 12GB of RAM:

```
$ /bin/time go test -run xxx -bench 'Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=5001' ./tsdb/
signal: killed
FAIL    github.com/prometheus/prometheus/tsdb   30.001s
FAIL
Command exited with non-zero status 1
35.45user 7.93system 0:30.86elapsed 140%CPU (0avgtext+0avgdata 7718416maxresident)k
```

after this change (look at `maxresident`):
```
$ /bin/time go test -run xxx -bench 'Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=5001' ./tsdb/
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
BenchmarkCompaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=5001-4                  1        3564781897 ns/op        284635744 B/op   6251195 allocs/op
PASS
ok      github.com/prometheus/prometheus/tsdb   79.468s
80.78user 7.93system 1:21.23elapsed 109%CPU (0avgtext+0avgdata 5831748maxresident)k
```


A smaller benchmark completes more quickly as well as running smaller:
before:
```
$ /bin/time go test -run xxx -bench 'Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=2001' ./tsdb/
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
BenchmarkCompaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=2001-4                  1        1755452710 ns/op        166213408 B/op   3251078 allocs/op
PASS
ok      github.com/prometheus/prometheus/tsdb   39.140s
45.24user 6.11system 0:39.87elapsed 128%CPU (0avgtext+0avgdata 3400696maxresident)k
```

after:
```
$ /bin/time go test -run xxx -bench 'Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=2001' ./tsdb/
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
BenchmarkCompaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=2001-4                  1        2111680952 ns/op        160974624 B/op   3251151 allocs/op
PASS
ok      github.com/prometheus/prometheus/tsdb   33.804s
34.32user 3.42system 0:34.56elapsed 109%CPU (0avgtext+0avgdata 2428068maxresident)k
```